docker/back/django_project/service/ocr/requirements.txt
Flask==1.1.1
pandas==0.25.3
requests==2.22.0
gunicorn==19.9.0
mecab-python3==0.7
Pillow==7.1.0
json5==0.8.5
numpy==1.17.4
jaconv==0.2.4
qrcode==6.1